+++ CHANGELOG.md
@@ -5,8 +5,7 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
-
-## **WORK IN PROGRESS**
+## 2.16.1 (2023-10-08)
 
 * (theimo1221) Shelly-Trv should respect HeatGroupSetting.automaticMode = false
 * (theimo1221) Update packages